requirements.txt
python-dotenv~=1.2.2
discord.py
dotenv~=0.9.9
jishaku
aiosqlite~=0.22.1
simpleeval~=1.0.7
mariadb~=1.1.14
aiohttp~=3.13.5